Change day 1 to use qi
Last year I tried to use `qi` (https://docs.racket-lang.org/qi/index.html)
for my puzzle solutions but found that it was a bit tricky to understand and
I seemed to struggle when working on a couple of puzzles, and I subsequently
gave up. This year I'm determined to use it and thankfully am able to start to
use it from the beginning, so that I can gain a better understanding of how the
library/DSL works on easy puzzles.

I've already gained quite a bit of understanding (that you need to think
in terms of values and lists, for example) and so I feel a bit more
confident going forward.

I'm curious how this will go as time goes on.